requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.6
scikit-learn
mlflow
pyaml
dagshub
dill
fastapi
uvicorn
python-multipart
-e .